[feat]change login to micro service

+++ web/controllers/AnswerController.go
@@ -95,9 +95,9 @@ func (this *AnswerController) GetUserProblems() {
 		this.Ctx.Redirect(304, "/index")
 		return
 	}
-	user_id := userSession.(int)
+	user_id := userSession.(int64)
 	pManage := initParticipantManage()
-	pReq := participantProto.PUserEventIdReq{EventId:int64(event_id),UserId:int64(user_id)}
+	pReq := participantProto.PUserEventIdReq{EventId:int64(event_id),UserId:user_id}
 	p,pErr := pManage.GetParticipantByUserAndEvent(context.TODO(),&pReq)
 	if pErr!=nil{
 		beego.Info("-------pErr--------",pErr)
@@ -195,9 +195,9 @@ func (this *AnswerController) GetUserAnswers() {
 		this.Ctx.Redirect(304, "/index")
 		return
 	}
-	user_id := userSession.(int)
+	user_id := userSession.(int64)
 	pManage := initParticipantManage()
-	pReq := participantProto.PUserEventIdReq{EventId:int64(event_id),UserId:int64(user_id)}
+	pReq := participantProto.PUserEventIdReq{EventId:int64(event_id),UserId:user_id}
 	p,pErr := pManage.GetParticipantByUserAndEvent(context.TODO(),&pReq)
 	if pErr!=nil{
 		beego.Info("-------pErr--------",pErr)
